Group drivers by adapter, not backing store
Closes #57
