--- event-store/src/adapters/amqp/mod.rs
@@ -1,5 +1,0 @@
-//! AQMP backed emitter adapter.
-
-mod emitter;
-
-pub use self::emitter::AMQPEmitterAdapter;
