--- event-store/src/adapters/redis/mod.rs
@@ -1,11 +1,0 @@
-mod cache;
-
-pub use self::cache::RedisCacheAdapter;
-
-use chrono::{DateTime, Utc};
-
-#[derive(Serialize, Deserialize)]
-struct RedisCacheItem<D> {
-    data: D,
-    time: DateTime<Utc>,
-}
