--- event-store/src/adapters/stub/mod.rs
@@ -1,5 +1,0 @@
-//! Stub implementations of things
-
-mod emitter;
-
-pub use self::emitter::StubEmitterAdapter;
