+++ event-store/src/prelude.rs
@@ -1,7 +1,6 @@
 //! Event store prelude
 
-pub use adapters::{CacheAdapter, CacheResult, EmitterAdapter, StoreAdapter};
+pub use adapters::{CacheAdapter, EmitterAdapter, StoreAdapter, StoreQuery};
 pub use aggregator::Aggregator;
 pub use event_store_derive_internals::{EventData, Events};
 pub use store::Store;
-pub use store_query::StoreQuery;
